requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

git+https://github.com/DemocracyClub/dc_signup_form.git@2.0.0
-e git+https://github.com/DemocracyClub/dc_base_theme.git@0.3.9#egg=dc_base_theme
-e git+https://github.com/DemocracyClub/django-hermes.git#egg=hermes
-e git+https://github.com/klen/django_markdown.git#egg=django_markdown
git+https://github.com/chrisdrackett/django-typogrify.git
beautifulsoup4==4.6.0
django-allauth==0.24.1
django-braces==1.12.0
django-cors-headers==1.1.0
django-extensions==1.7.6
django-localflavor==1.1
django-markdown-deux==1.0.5
django-model-utils==2.3.1
django-ratelimit==0.6.0
Django==1.10.8  # pyup: >=1.11,<2.0
djangorestframework==3.5.4
gocardless_pro==1.1.0
jsonfield==1.0.3
markdown==2.6.2
Pillow==4.3.0
plotly==1.9.10
psycopg2-binary==2.7.4

sorl-thumbnail==12.3


django-storages==1.6.5

boto3==1.7.74

sentry-sdk==0.14.3

black==19.10b0
